Code review improvements
